requirements.txt
# Google Cloud Functions framework and Storage client
functions-framework==3.*
google-cloud-storage

# For video and image processing
opencv-python-headless
ffmpeg-python

# The AI and ML frameworks
torch
torchvision
# Note: We will add the specific Real-ESRGAN library later